server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.12
sqlalchemy==2.0.36
geoalchemy2==0.16.0
alembic==1.14.0
asyncpg==0.30.0
python-dotenv==1.0.1
requests==2.32.3
httpx==0.28.1
geoip2==4.8.1
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
shapely==2.0.6
